// projects/codx-ac/src/lib/models/IVPostingAccounts.model.ts
export class IVPostingAccounts {
  accountID: any;
  allowEditKey: any;
  assign: any;
  buid: any;
  createdBy: any;
  createdOn: any;
  currencyID: any;
  custLevel: any;
  custSelection: any;
  custType: any;
  delete: any;
  diM1: any;
  diM1Control: any;
  diM2: any;
  diM2Control: any;
  diM3: any;
  diM3Control: any;
  idField: any;
  includeTables: any;
  itemLevel: any;
  itemSelection: any;
  modifiedBy: any;
  modifiedOn: any;
  moduleID: any;
  note: any;
  orderPoolID: any;
  postType: any;
  prodStageID: any;
  projectID: any;
  recID: any;
  share: any;
  sorting: any;
  unbounds: any;
  updateColumns: any;
  useCustomer: any;
  warehouseID: any;
  write: any;
}